Code Quality: rewrite codes, move join() from zk.py to joint.py
